When FetcherRunnable detects about outOfRange, it tries to fix the issue by bump up both fectedoffset and consumedoffset ahead to the valid one. As Fetcher thread don't own and shouldn't own the consumedoffset, the consumerIterator who is the actual owner of consumedoffset update the consumedoffset back to the position it actually consumed. In the next iteration of Fetcher thread, it detects that the lag between fectedoffset consumedoffset is big enough so it stop fetching more. It is true during normal case, but it is not true when the fetchedoffset being manually set ahead instead of fetched ahead.
The fix is first, FetcherRunnable shouldn't just set the consumedoffset as it doesn't own how far it consumed. When FetcherRunnable detects about outofrange error, it signal PartitionTopicInfo to fix its offset. Creae a private variable lastKnownGoodNextRequestOffset for pti to save the old marker of NextRequestOffset When OutOfRange happens. In the next iteration of fetcher, use that to caculate the actual messages in the queue. After the successful fetch, it set back to align with NextRequestOffset.

Also cleaned up messageOffset as it is never being used.